requirements.txt
Flask
pandas
numpy
scikit-learn
tensorflow-cpu
